spark-streaming-pipeline/requirements.txt
# Spark Streaming and PySpark
pyspark==3.5.7

# Kafka for Python (consumer/producer)
kafka-python==2.1.1
confluent-kafka==2.3.0  # Added: Better performance, official client

# Data processing and manipulation
pandas==2.1.1
numpy==1.26.0
pyarrow==14.0.1  # Added: Better Spark-Pandas integration

# Visualization
matplotlib==3.8.0
seaborn==0.12.2
plotly==5.17.0  # Added: Interactive Grafana-like charts

# Date/time handling
python-dateutil==2.9.2
pytz==2023.3.post1  # Added: Timezone support for timestamps

# JSON validation and processing
jsonschema==4.21.0
ujson==5.8.0  # Added: Faster JSON parsing for streaming

# Advanced logging
loguru==0.7.0

# Docker client
docker==6.1.3

# Database connectors
psycopg2-binary==2.9.9
mysql-connector-python==8.1.1
sqlalchemy==2.0.23  # Added: ORM for database operations

# Machine Learning (for your ML models)
scikit-learn==1.3.2  # Added: Required for Spark ML interoperability
joblib==1.3.2  # Added: Model serialization

# Web/API (for monitoring/health checks)
requests==2.31.0  # Added: API calls for service health checks
fastapi==0.104.1  # Added: Optional API for predictions
uvicorn==0.24.0  # Added: ASGI server

# Utilities
python-dotenv==1.0.0  # Added: Environment variable management
tqdm==4.66.1  # Added: Progress bars for data processing
click==8.1.7  # Added: Command-line interfaces

# Testing
pytest==7.4.3  # Added: Unit tests for your pipeline
pytest-asyncio==0.21.1  # Added: Async tests

# Performance
numba==0.58.1  # Added: JIT compilation for data processing